refactor: WinningNumbers 생성자 변경에 따른 테스트 수정
