--- src/main/java/lotto/validation/InputValidator.java
@@ -1,21 +1,0 @@
-package lotto.validation;
-
-/**
- * 테스트에서만 사용됨
- */
-public class InputValidator {
-
-    public static void validateIsBlank(String input) {
-        if (input == null || input.isBlank()) {
-            throw new IllegalArgumentException("[ERROR] 입력값이 비어있습니다.");
-        }
-    }
-
-    public static void validateIsInteger(String input) {
-        try {
-            Integer.parseInt(input);
-        } catch (NumberFormatException e) {
-            throw new IllegalArgumentException("[ERROR] 입력값이 Integer 타입이 아닙니다.");
-        }
-    }
-}
