+++ src/test/java/lotto/model/LottoServiceTest.java
@@ -15,7 +15,7 @@ class LottoServiceTest {
     void 보너스_번호가_당첨_번호들과_중복되지_않으면_테스트에_성공한다() {
         // given
         BonusNumber bonusNumber = new BonusNumber(7);
-        WinningNumbers winningNumbers = new WinningNumbers(List.of("1", "2", "3", "4", "5", "6"));
+        WinningNumbers winningNumbers = new WinningNumbers(List.of(1, 2, 3, 4, 5, 6));
 
         // when & then
         assertThatCode(() -> lottoService.checkDuplicate(winningNumbers, bonusNumber))
@@ -26,7 +26,7 @@ class LottoServiceTest {
     void 보너스_번호가_당첨_번호들과_중복되면_예외가_발생한다() {
         // given
         BonusNumber bonusNumber = new BonusNumber(3);
-        WinningNumbers winningNumbers = new WinningNumbers(List.of("1", "2", "3", "4", "5", "6"));
+        WinningNumbers winningNumbers = new WinningNumbers(List.of(1, 2, 3, 4, 5, 6));
 
         // when & then
         assertThatThrownBy(() -> lottoService.checkDuplicate(winningNumbers, bonusNumber))
